requirements.txt
cryptography==42.0.8
pyperclip==1.9.0
python-dotenv==1.0.1
python_bcrypt==0.3.2
